refactor: decompose Register and separate to components
